# documentation/openapi.yaml
openapi: 3.0.0
info:
  title: Projects
  contact: {}
  version: '1.0'
servers:
  - url: http://example.com
    variables: {}
paths:
  /projects:
    post:
      security:
      - bearer: []
      tags:
        - Project
      summary: CREATE project
      operationId: CREATEproject
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CREATEprojectRequest'
            example:
              name: nombre2
              description: algo 123
              status: true
              users: hola
        required: true
      responses:
        '200':
          description: OK
          headers:
            X-Powered-By:
              content:
                text/plain:
                  schema:
                    type: string
                    example: Express
                  example: Express
            Access-Control-Allow-Origin:
              content:
                text/plain:
                  schema:
                    type: string
                    example: '*'
                  example: '*'
            Content-Length:
              content:
                text/plain:
                  schema:
                    type: string
                    example: '517'
                  example: '517'
            ETag:
              content:
                text/plain:
                  schema:
                    type: string
                    example: W/"205-FmXWzkEztYkw+mZCLGXxLFOgtl4"
                  example: W/"205-FmXWzkEztYkw+mZCLGXxLFOgtl4"
            Date:
              content:
                text/plain:
                  schema:
                    type: string
                    example: Sat, 09 Apr 2022 01:49:49 GMT
                  example: Sat, 09 Apr 2022 01:49:49 GMT
            Connection:
              content:
                text/plain:
                  schema:
                    type: string
                    example: keep-alive
                  example: keep-alive
            Keep-Alive:
              content:
                text/plain:
                  schema:
                    type: string
                    example: timeout=5
                  example: timeout=5
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/m200'
              example:
                meta:
                  status: 200
                  errors: null
                data:
                  dataValues:
                    id: 25
                    name: nombre
                    description: algo 123
                    status: true
                    updatedAt: '2022-04-09T01:49:49.764Z'
                    createdAt: '2022-04-09T01:49:49.764Z'
                  _previousDataValues:
                    name: nombre
                    description: algo 123
                    status: true
                    id: 25
                    createdAt: '2022-04-09T01:49:49.764Z'
                    updatedAt: '2022-04-09T01:49:49.764Z'
                  uniqno: 1
                  _changed: {}
                  _options:
                    isNewRecord: true
                    _schema: null
                    _schemaDelimiter: ''
                  isNewRecord: false
                msg: Project created successfuly
        '401':
          description: Unauthorized
          headers:
            X-Powered-By:
              content:
                text/plain:
                  schema:
                    type: string
                    example: Express
                  example: Express
            Access-Control-Allow-Origin:
              content:
                text/plain:
                  schema:
                    type: string
                    example: '*'
                  example: '*'
            Content-Length:
              content:
                text/plain:
                  schema:
                    type: string
                    example: '99'
                  example: '99'
            ETag:
              content:
                text/plain:
                  schema:
                    type: string
                    example: W/"63-bW21xeFuOlW1sz6BLIPj/BfTLHA"
                  example: W/"63-bW21xeFuOlW1sz6BLIPj/BfTLHA"
            Date:
              content:
                text/plain:
                  schema:
                    type: string
                    example: Sat, 09 Apr 2022 17:02:02 GMT
                  example: Sat, 09 Apr 2022 17:02:02 GMT
            Connection:
              content:
                text/plain:
                  schema:
                    type: string
                    example: keep-alive
                  example: keep-alive
            Keep-Alive:
              content:
                text/plain:
                  schema:
                    type: string
                    example: timeout=5
                  example: timeout=5
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/401Nopermission1'
              example:
                meta:
                  status: 401
                  ok: false
                errors:
                  msg: You do not have permissions for this resource
        '403':
          description: Forbidden
          headers:
            X-Powered-By:
              content:
                text/plain:
                  schema:
                    type: string
                    example: Express
                  example: Express
            Access-Control-Allow-Origin:
              content:
                text/plain:
                  schema:
                    type: string
                    example: '*'
                  example: '*'
            Content-Length:
              content:
                text/plain:
                  schema:
                    type: string
                    example: '50'
                  example: '50'
            ETag:
              content:
                text/plain:
                  schema:
                    type: string
                    example: W/"32-zKXBNALrhPJLhGrNy32aZ4MX1MI"
                  example: W/"32-zKXBNALrhPJLhGrNy32aZ4MX1MI"
            Date:
              content:
                text/plain:
                  schema:
                    type: string
                    example: Sat, 09 Apr 2022 17:13:04 GMT
                  example: Sat, 09 Apr 2022 17:13:04 GMT
            Connection:
              content:
                text/plain:
                  schema:
                    type: string
                    example: keep-alive
                  example: keep-alive
            Keep-Alive:
              content:
                text/plain:
                  schema:
                    type: string
                    example: timeout=5
                  example: timeout=5
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/403Tokenrequired1'
              example:
                error: A token is required for authentication
        '422':
          description: Unprocessable Entity
          headers:
            X-Powered-By:
              content:
                text/plain:
                  schema:
                    type: string
                    example: Express
                  example: Express
            Access-Control-Allow-Origin:
              content:
                text/plain:
                  schema:
                    type: string
                    example: '*'
                  example: '*'
            Content-Length:
              content:
                text/plain:
                  schema:
                    type: string
                    example: '77'
                  example: '77'
            ETag:
              content:
                text/plain:
                  schema:
                    type: string
                    example: W/"4d-BZvTAKjzk6pgSnqRQTMV57Spihc"
                  example: W/"4d-BZvTAKjzk6pgSnqRQTMV57Spihc"
            Date:
              content:
                text/plain:
                  schema:
                    type: string
                    example: Sat, 09 Apr 2022 01:50:23 GMT
                  example: Sat, 09 Apr 2022 01:50:23 GMT
            Connection:
              content:
                text/plain:
                  schema:
                    type: string
                    example: keep-alive
                  example: keep-alive
            Keep-Alive:
              content:
                text/plain:
                  schema:
                    type: string
                    example: timeout=5
                  example: timeout=5
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/422Inputrequired1'
              example:
                errors:
                  - msg: Input required
                    param: description
                    location: body
        '500':
          description: Internal Server Error
          headers:
            X-Powered-By:
              content:
                text/plain:
                  schema:
                    type: string
                    example: Express
                  example: Express
            Access-Control-Allow-Origin:
              content:
                text/plain:
                  schema:
                    type: string
                    example: '*'
                  example: '*'
            Content-Length:
              content:
                text/plain:
                  schema:
                    type: string
                    example: '35'
                  example: '35'
            ETag:
              content:
                text/plain:
                  schema:
                    type: string
                    example: W/"23-PmmVM4t6OFE2CKfH8e8hjD7fvBc"
                  example: W/"23-PmmVM4t6OFE2CKfH8e8hjD7fvBc"
            Date:
              content:
                text/plain:
                  schema:
                    type: string
                    example: Sat, 09 Apr 2022 01:00:23 GMT
                  example: Sat, 09 Apr 2022 01:00:23 GMT
            Connection:
              content:
                text/plain:
                  schema:
                    type: string
                    example: keep-alive
                  example: keep-alive
            Keep-Alive:
              content:
                text/plain:
                  schema:
                    type: string
                    example: timeout=5
                  example: timeout=5
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/500ServerError1'
              example:
                meta:
                  status: 500
                  errors: {}
      deprecated: false
    get:
      security:
      - bearer: []
      tags:
        - Project
      summary: GET projects by name or page
      operationId: GETprojectsbynameorpage
      parameters:
        - name: page
          in: query
          description: ''
          required: true
          style: form
          explode: true
          schema:
            type: integer
            format: int32
            example: 1
      responses:
        '200':
          description: OK
          headers:
            X-Powered-By:
              content:
                text/plain:
                  schema:
                    type: string
                    example: Express
                  example: Express
            Access-Control-Allow-Origin:
              content:
                text/plain:
                  schema:
                    type: string
                    example: '*'
                  example: '*'
            Content-Length:
              content:
                text/plain:
                  schema:
                    type: string
                    example: '1853'
                  example: '1853'
            ETag:
              content:
                text/plain:
                  schema:
                    type: string
                    example: W/"73d-lQpxxtBMXEostdTeiL7Q4+CPTJc"
                  example: W/"73d-lQpxxtBMXEostdTeiL7Q4+CPTJc"
            Date:
              content:
                text/plain:
                  schema:
                    type: string
                    example: Sat, 09 Apr 2022 02:35:12 GMT
                  example: Sat, 09 Apr 2022 02:35:12 GMT
            Connection:
              content:
                text/plain:
                  schema:
                    type: string
                    example: keep-alive
                  example: keep-alive
            Keep-Alive:
              content:
                text/plain:
                  schema:
                    type: string
                    example: timeout=5
                  example: timeout=5
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/200Pagequery'
              example:
                prev: null
                next: http://undefined/projects?page=2
                results:
                  - id: 1
                    name: new Name
                    description: New description
                    status: true
                    projectManagerId: 2
                    createdAt: '2022-04-08T20:07:18.000Z'
                    updatedAt: '2022-04-09T01:53:33.000Z'
                  - id: 4
                    name: Project 1
                    description: Project 1 description
                    status: true
                    projectManagerId: 1
                    createdAt: '2022-04-08T20:11:37.000Z'
                    updatedAt: '2022-04-08T20:11:37.000Z'
                  - id: 5
                    name: Project 2
                    description: Project 2 description
                    status: true
                    projectManagerId: 1
                    createdAt: '2022-04-08T20:11:37.000Z'
                    updatedAt: '2022-04-08T20:11:37.000Z'
                  - id: 6
                    name: Project 3
                    description: Project 3 description
                    status: true
                    projectManagerId: 1
                    createdAt: '2022-04-08T20:11:37.000Z'
                    updatedAt: '2022-04-08T20:11:37.000Z'
                  - id: 7
                    name: Project 4
                    description: Project 4 description
                    status: true
                    projectManagerId: 1
                    createdAt: '2022-04-08T20:11:37.000Z'
                    updatedAt: '2022-04-08T20:11:37.000Z'
                  - id: 8
                    name: Project 5
                    description: Project 5 description
                    status: true
                    projectManagerId: 1
                    createdAt: '2022-04-08T20:11:37.000Z'
                    updatedAt: '2022-04-08T20:11:37.000Z'
                  - id: 9
                    name: Project 6
                    description: Project 6 description
                    status: true
                    projectManagerId: 1
                    createdAt: '2022-04-08T20:11:37.000Z'
                    updatedAt: '2022-04-08T20:11:37.000Z'
                  - id: 10
                    name: Project 7
                    description: Project 7 description
                    status: true
                    projectManagerId: 1
                    createdAt: '2022-04-08T20:11:37.000Z'
                    updatedAt: '2022-04-08T20:11:37.000Z'
                  - id: 11
                    name: Project 8
                    description: Project 8 description
                    status: true
                    projectManagerId: 1
                    createdAt: '2022-04-08T20:11:37.000Z'
                    updatedAt: '2022-04-08T20:11:37.000Z'
                  - id: 12
                    name: Project 9
                    description: Project 9 description
                    status: true
                    projectManagerId: 1
                    createdAt: '2022-04-08T20:11:37.000Z'
                    updatedAt: '2022-04-08T20:11:37.000Z'
        '403':
          description: Forbidden
          headers:
            X-Powered-By:
              content:
                text/plain:
                  schema:
                    type: string
                    example: Express
                  example: Express
            Access-Control-Allow-Origin:
              content:
                text/plain:
                  schema:
                    type: string
                    example: '*'
                  example: '*'
            Content-Length:
              content:
                text/plain:
                  schema:
                    type: string
                    example: '50'
                  example: '50'
            ETag:
              content:
                text/plain:
                  schema:
                    type: string
                    example: W/"32-zKXBNALrhPJLhGrNy32aZ4MX1MI"
                  example: W/"32-zKXBNALrhPJLhGrNy32aZ4MX1MI"
            Date:
              content:
                text/plain:
                  schema:
                    type: string
                    example: Sat, 09 Apr 2022 17:17:34 GMT
                  example: Sat, 09 Apr 2022 17:17:34 GMT
            Connection:
              content:
                text/plain:
                  schema:
                    type: string
                    example: keep-alive
                  example: keep-alive
            Keep-Alive:
              content:
                text/plain:
                  schema:
                    type: string
                    example: timeout=5
                  example: timeout=5
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/403Tokenrequired1'
              example:
                error: A token is required for authentication
        '500':
          description: Internal Server Error
          headers:
            X-Powered-By:
              content:
                text/plain:
                  schema:
                    type: string
                    example: Express
                  example: Express
            Access-Control-Allow-Origin:
              content:
                text/plain:
                  schema:
                    type: string
                    example: '*'
                  example: '*'
            Content-Length:
              content:
                text/plain:
                  schema:
                    type: string
                    example: '35'
                  example: '35'
            ETag:
              content:
                text/plain:
                  schema:
                    type: string
                    example: W/"23-PmmVM4t6OFE2CKfH8e8hjD7fvBc"
                  example: W/"23-PmmVM4t6OFE2CKfH8e8hjD7fvBc"
            Date:
              content:
                text/plain:
                  schema:
                    type: string
                    example: Sat, 09 Apr 2022 01:57:55 GMT
                  example: Sat, 09 Apr 2022 01:57:55 GMT
            Connection:
              content:
                text/plain:
                  schema:
                    type: string
                    example: keep-alive
                  example: keep-alive
            Keep-Alive:
              content:
                text/plain:
                  schema:
                    type: string
                    example: timeout=5
                  example: timeout=5
          content:
            application/json; charset=utf-8:
              schema:
                $ref: '#/components/schemas/500ServerError1'
              example:
                meta:
                  status: 500
                  errors: {}
      deprecated: false
...
components:
  schemas:
    CREATEprojectRequest:
      title: CREATEprojectRequest
      required:
        - name
        - description
        - status
        - users
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: boolean
        users:
          type: string
      example:
        name: nombre2
        description: algo 123
        status: true
        users: hola
    m200:
      title: m200
      required:
        - meta
        - data
        - msg
      type: object
      properties:
        meta:
          $ref: '#/components/schemas/Meta'
        data:
          $ref: '#/components/schemas/Data'
        msg:
          type: string
      example:
        meta:
          status: 200
          errors: null
        data:
          dataValues:
            id: 25
            name: nombre
            description: algo 123
            status: true
            updatedAt: '2022-04-09T01:49:49.764Z'
            createdAt: '2022-04-09T01:49:49.764Z'
          _previousDataValues:
            name: nombre
            description: algo 123
            status: true
            id: 25
            createdAt: '2022-04-09T01:49:49.764Z'
            updatedAt: '2022-04-09T01:49:49.764Z'
          uniqno: 1
          _changed: {}
          _options:
            isNewRecord: true
            _schema: null
            _schemaDelimiter: ''
          isNewRecord: false
        msg: Project created successfuly
    Meta:
      title: Meta
      required:
        - status
        - errors
      type: object
      properties:
        status:
          type: integer
          format: int32
        errors:
          type: string
          nullable: true
      example:
        status: 200
        errors: null
    Data:
      title: Data
      required:
        - dataValues
        - _previousDataValues
        - uniqno
        - _changed
        - _options
        - isNewRecord
      type: object
      properties:
        dataValues:
          $ref: '#/components/schemas/DataValues'
        _previousDataValues:
          $ref: '#/components/schemas/PreviousDataValues'
        uniqno:
          type: integer
          format: int32
        _changed:
          type: object
        _options:
          $ref: '#/components/schemas/Options'
        isNewRecord:
          type: boolean
      example:
        dataValues:
          id: 25
          name: nombre
          description: algo 123
          status: true
          updatedAt: '2022-04-09T01:49:49.764Z'
          createdAt: '2022-04-09T01:49:49.764Z'
        _previousDataValues:
          name: nombre
          description: algo 123
          status: true
          id: 25
          createdAt: '2022-04-09T01:49:49.764Z'
          updatedAt: '2022-04-09T01:49:49.764Z'
        uniqno: 1
        _changed: {}
        _options:
          isNewRecord: true
          _schema: null
          _schemaDelimiter: ''
        isNewRecord: false
    DataValues:
      title: DataValues
      required:
        - id
        - name
        - description
        - status
        - updatedAt
        - createdAt
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        description:
          type: string
        status:
          type: boolean
        updatedAt:
          type: string
        createdAt:
          type: string
      example:
        id: 25
        name: nombre
        description: algo 123
        status: true
        updatedAt: '2022-04-09T01:49:49.764Z'
        createdAt: '2022-04-09T01:49:49.764Z'
    PreviousDataValues:
      title: PreviousDataValues
      required:
        - name
        - description
        - status
        - id
        - createdAt
        - updatedAt
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: boolean
        id:
          type: integer
          format: int32
        createdAt:
          type: string
        updatedAt:
          type: string
      example:
        name: nombre
        description: algo 123
        status: true
        id: 25
        createdAt: '2022-04-09T01:49:49.764Z'
        updatedAt: '2022-04-09T01:49:49.764Z'
    Options:
      title: Options
      required:
        - isNewRecord
        - _schema
        - _schemaDelimiter
      type: object
      properties:
        isNewRecord:
          type: boolean
        _schema:
          type: string
          nullable: true
        _schemaDelimiter:
          type: string
      example:
        isNewRecord: true
        _schema: null
        _schemaDelimiter: ''
    Meta1:
      title: Meta1
      required:
        - status
        - ok
      type: object
      properties:
        status:
          type: integer
          format: int32
        ok:
          type: boolean
      example:
        status: 401
        ok: false
    Errors:
      title: Errors
      required:
        - msg
      type: object
      properties:
        msg:
          type: string
      example:
        msg: You do not have permissions for this resource
    401Nopermission1:
      title: 401Nopermission1
      required:
        - meta
        - errors
      type: object
      properties:
        meta:
          $ref: '#/components/schemas/Meta1'
        errors:
          $ref: '#/components/schemas/Errors'
      example:
        meta:
          status: 401
          ok: false
        errors:
          msg: You do not have permissions for this resource
    Errors2:
      title: Errors2
      required:
        - msg
        - param
        - location
      type: object
      properties:
        msg:
          type: string
        param:
          type: string
        location:
          type: string
      example:
        msg: Input required
        param: description
        location: body
    422Inputrequired1:
      title: 422Inputrequired1
      required:
        - errors
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/Errors2'
          description: ''
      example:
        errors:
          - msg: Input required
            param: description
            location: body
    403Tokenrequired1:
      title: 403Tokenrequired1
      required:
        - error
      type: object
      properties:
        error:
          type: string
      example:
        error: A token is required for authentication
    Meta3:
      title: Meta3
      required:
        - status
        - errors
      type: object
      properties:
        status:
          type: integer
          format: int32
        errors:
          type: object
      example:
        status: 500
        errors: {}
    500ServerError1:
      title: 500ServerError1
      required:
        - meta
      type: object
      properties:
        meta:
          $ref: '#/components/schemas/Meta3'
      example:
        meta:
          status: 500
          errors: {}
...
    200Pagequery:
      title: 200Pagequery
      required:
        - prev
        - next
        - results
      type: object
      properties:
        prev:
          type: string
          nullable: true
        next:
          type: string
        results:
          type: array
          items:
            $ref: '#/components/schemas/Result'
          description: ''
      example:
        prev: null
        next: http://undefined/projects?page=2
        results:
          - id: 1
            name: new Name
            description: New description
            status: true
            projectManagerId: 2
            createdAt: '2022-04-08T20:07:18.000Z'
            updatedAt: '2022-04-09T01:53:33.000Z'
          - id: 4
            name: Project 1
            description: Project 1 description
            status: true
            projectManagerId: 1
            createdAt: '2022-04-08T20:11:37.000Z'
            updatedAt: '2022-04-08T20:11:37.000Z'
          - id: 5
            name: Project 2
            description: Project 2 description
            status: true
            projectManagerId: 1
            createdAt: '2022-04-08T20:11:37.000Z'
            updatedAt: '2022-04-08T20:11:37.000Z'
          - id: 6
            name: Project 3
            description: Project 3 description
            status: true
            projectManagerId: 1
            createdAt: '2022-04-08T20:11:37.000Z'
            updatedAt: '2022-04-08T20:11:37.000Z'
          - id: 7
            name: Project 4
            description: Project 4 description
            status: true
            projectManagerId: 1
            createdAt: '2022-04-08T20:11:37.000Z'
            updatedAt: '2022-04-08T20:11:37.000Z'
          - id: 8
            name: Project 5
            description: Project 5 description
            status: true
            projectManagerId: 1
            createdAt: '2022-04-08T20:11:37.000Z'
            updatedAt: '2022-04-08T20:11:37.000Z'
          - id: 9
            name: Project 6
            description: Project 6 description
            status: true
            projectManagerId: 1
            createdAt: '2022-04-08T20:11:37.000Z'
            updatedAt: '2022-04-08T20:11:37.000Z'
          - id: 10
            name: Project 7
            description: Project 7 description
            status: true
            projectManagerId: 1
            createdAt: '2022-04-08T20:11:37.000Z'
            updatedAt: '2022-04-08T20:11:37.000Z'
          - id: 11
            name: Project 8
            description: Project 8 description
            status: true
            projectManagerId: 1
            createdAt: '2022-04-08T20:11:37.000Z'
            updatedAt: '2022-04-08T20:11:37.000Z'
          - id: 12
            name: Project 9
            description: Project 9 description
            status: true
            projectManagerId: 1
            createdAt: '2022-04-08T20:11:37.000Z'
            updatedAt: '2022-04-08T20:11:37.000Z'
    Result:
      title: Result
      required:
        - id
        - name
        - description
        - status
        - projectManagerId
        - createdAt
        - updatedAt
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        description:
          type: string
        status:
          type: boolean
        projectManagerId:
          type: integer
          format: int32
        createdAt:
          type: string
        updatedAt:
          type: string
      example:
        id: 1
        name: new Name
        description: New description
        status: true
        projectManagerId: 2
        createdAt: '2022-04-08T20:07:18.000Z'
        updatedAt: '2022-04-09T01:53:33.000Z'
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security:
  - bearer: []
...